added executor logic

+++ executor/utils.js
@@ -3,18 +3,18 @@ const exec = util.promisify(require("child_process").exec);
 const fs = require("fs");
 
 module.exports.execute = async (command, cleanupFiles) => {
-  let result;
+  let result = { stdout: "", stderr: "" };
   try {
-    result.stdout = await exec(command, {
+    result = await exec(command, {
       timeout: 1000,
       maxBuffer: 1024 * 1024 * 5,
     });
   } catch (err) {
-    result.stderr = err;
+    result = err;
   } finally {
     await Promise.all(
       cleanupFiles.map(async (file) => {
-        await fs.promises.unlink(file);
+        if (fs.existsSync(file)) await fs.promises.unlink(file);
       })
     );
   }
